Handle check_run completions with ci staging
Right now we have "three" stages:
  1. non-monorepo: none (schedule everything)
  2. mono-repo: engine, tests.

To handle 2-engine, we need to react to updates from check_run events.
This PR only cares about flutter/flaux right now, but it will care about
any monorepo in the future. The engine check is hard-coded because to
put it somewhere else would be to either create a new firestore doc or
encode it in the tree (e.g. .ci.yaml) - but we're not interested in
building a build-graph right now.