[mq] delete temp branches when PR is actually merged (#4043)
In the merge queue world it is incorrect to assume that a PR is immediately merged when it's submitted. The PR may be in the queue for a while. Deleting the source branch leads to PR being [removed from the queue and closed](https://github.com/flutter/flutter/issues/159365).

This PR moves the logic that deletes the branch from the revert validation hook to the `pull_request/closed` hook.

Fixes https://github.com/flutter/flutter/issues/159365