remove dangerously unused SHA from parsed MQ branch (#4138)
The SHA in the branch name could be accidentally used, and this logic's presence already confused me while I was investigating https://github.com/flutter/flutter/issues/160704. The branch name is not meant to be the source of truth for the actual head SHA that's being tested (that should be the `headSha` of the `merge_group` request). In the spirit of "better safe than sorry", let's remove it for now.