Fix exemption for flutter tool Kotlin tests, add test, and update README (#4169)
Follow up to https://github.com/flutter/cocoon/pull/4162.

The first PR was wrong - it modified the logic that is used to handle packages PRs, not framework PRs. This PR fixes that, adds a test, and also modifies the README to add a missing step that I found in https://github.com/flutter/flutter/issues/129651, which originally blocked me from running the tests locally.