Revert "Use GraphQL to get the commit timestamp and parent sha  (#4218)" (#4228)
This reverts commit c71f46eb517faac97a567a9474db8217e5564be1.

GraphQL is returning a 500 error when searching fro mergedAtCommit; this
is not handled as a retry and we fail to schedule

@jtmcdole:

> > Unable to query committedDate

> I think we need to roll back cocoon or revert my change and roll
forward. This is the code that is trying to find the parent sha when the
PR is closed (and merged) - the push message does not appear to be
retried and we're not scheduling post submits.
>
> We should have known about this earlier, but cocoon not rolling out
let it slip