Use `master` as both the channel and branch name. (#4300)
Closes https://github.com/flutter/flutter/issues/164726.

It turns out `HEAD`, as defined, did nothing, and was really just ignored (by both the UI and backend), so trying to make it do something broke stuff. I thought about adding a UI regression test, but there isn't (really) a way to know it would work (or not), so settled for a note in a backend regression test.