Reword the missing test bot comment (#4305)
Currently, this message is often the first thing new contributors (who are not familiar with our processes) see, telling them to ask for a test exemption in Discord. This results in a lot of people asking for exemptions on PRs that are clearly not exempt, and that any reviewer would tell them to add tests for, putting an unnecessary burden on the small number of exemption reviewers.

This moves the section about requesting an exemption to the paragraph targeted for reviewers, to try to encourage the desired outcome of having exemption requests happen only once reviewers have agreed that it should be exempt.