Move `SecretManager` to `package/cocoon_server[_test]`. (#4372)
Towards https://github.com/flutter/flutter/issues/142951.

Datastore is currently used (incorrectly) to store secrets in `app_dart`. This refactor moves (a refactored version of) the `SecretManager` in `auto_submit` to `cocoon_server[_test]`, in preparation for it to be used in `app_dart` in a follow-up PR (and after moving secrets).