Start using `SecretManager` in `app_dart` (#4373)
Changes the use of datastore for application secrets to `SecretManager`.

In addition, moves `bin/local_server.dart` to `tool/local_server.dart`. This is the pattern that should have been used all along - it allows using test (dev) dependencies, which appears to be the intention, without making the test implementations non-dev API.

All of the secrets were copied from datastore -> secret manager.

If this PR lands and sticks, we can remove the datastore storage.