Reduce amount of errors/warnings that reoccur due to web hooks. (#4393)
A 5xx class error returned to Google Cloud PubSub basically means "retry this forever", which is unfortunate, because for some classes of errors we return today, we will likely never succeed until the heat death of the universe. This PR adds a new `404`-based error response (4xx class will not be retried) in order to tamp these down.