Implement fakes of `FirestoreService` except `documentsResource()` (#4443)
Closes https://github.com/flutter/flutter/issues/165931.

A basic in-memory implementation of Firestore in-so-far it supports `FirestoreService`.

For more advanced use-cases, i.e. `documentsResource().any.possible.api()`, mocking is still required, but for much of our API surface, it now _really_ reads, writes, and updates to an in-memory database that supports queries, filtering, ordering, and a (very basic) transaction.

We can move more code from `.documentsResource()` to `service.something()`, and ensure those are supported by the fake. In a follow-up PR we can start using this in our own tests where appropriate.

No production code changes.