Rewrite `vacuum-stale-tasks` to be Firestore first. (#4446)
This PR refactors `VaccumStaleTasks` to use Firestore as the source of truth (instead of Datastore), and makes the updates to Datastore secondary instead of primary.  I tried to structure it in such a way that removing the datastore code would be easy.

Builds on https://github.com/flutter/cocoon/pull/4443 by adding `firestore_matchers.dart`, which are a nice way to make expectations against `FakeFirestoreService`, including the matcher `existsInStorage`.

Made a few small bug fixes to `FakeFirestoreService` along the way.