Use a Firestore-first model for `dart_internal_subscription`. (#4449)
Similar to https://github.com/flutter/cocoon/pull/4446.

This PR refactors `DartInternalSubscription` to use Firestore as the source of truth (instead of Datastore), and makes the updates to Datastore secondary instead of primary. I tried to structure it in such a way that removing the datastore code would be easy.