Refactor `GerritService`, use retries for `GET`s, fix FIXME (#4468)
This PR removes a FIXME, lints for FIXMEs in the future, and fixes the test.

The original test did nothing, because if you write:
```dart
try {
  doSomethingThatNeverThrows();
} catch (e) {
  expect(e, 'You can literally write anything here');
}
```

The test will always pass. It turns out `GerritService` never retried (or threw) on `.get()` requests.

Fixing this and using a retry-client will also reduce the amount of pub-sub messages we use just to check Gerrit when waiting a few seconds in thread would have been sufficient. In addition did some TLC to `GerritService` and `FakeGerritService` across the repo.