Treat `rerun-prod-task?task=all` as marking all statuses as new. (#4566)
Closes https://github.com/flutter/flutter/issues/165816.

Since we have several hundred tasks, scheduling them all in an RPC is not viable. This PR repurposes `rerun-prod-task?task=all` (the RPC powering the button labeled below) to mark all tasks as `New`, so that the _backfiller_ picks them up. We can change priorities (i.e. have the backfiller fill release candidates with priority) in a future PR if desired.

Also changes the branch to run the query/updates in a transaction.

![Screenshot 2025-04-21 at 10 52 52 AM](https://github.com/user-attachments/assets/e2b9e3fa-7491-47c4-aaba-c142ef25a0d3)

- Towards https://github.com/flutter/flutter/issues/163896.
- Unblocks https://github.com/flutter/flutter/issues/165078.
- Unblocks https://github.com/flutter/flutter/issues/167383.