Add and use a `TimeRange` class to represent before/after timestamp. (#4719)
I started to work on https://github.com/flutter/flutter/issues/169234 and realized the current way timestamp filtering is done with Firebase is inflexible (and personally, confusing). This adds a small-ish `TimeRange` class in `cocoon_commons` and uses it instead of `int? timestamp` to determine which commits to find.

(Even if we want a different approach than https://github.com/flutter/flutter/issues/169234, this seems worth it?)