Update "Settings" popup to grow in size vertically (#4789)
TIL the "Settings" popup (triggered by clicking the gear button) has buttons at the bottom but they are hidden by the fixed height. Unfortunately on MacOS the scroll bars are hidden by default so I never knew they were there.
<img width="1179" alt="Screenshot 2025-06-24 at 10 49 30 AM" src="https://github.com/user-attachments/assets/51666a6e-dccd-47b4-8011-cf0fb3b20ecd" />

With this change the height can now expand enough to fix the contents.
<img width="1288" alt="Screenshot 2025-06-24 at 10 40 16 AM" src="https://github.com/user-attachments/assets/5b7cc3b4-dcd5-4f73-a6de-8ce3ee7d1ca7" />

There are some overflow issues on very small screen sizes so if there are preferable ways to achieve this please let me know. 
Ideally it should resize to fit the contents if the screen size allows, otherwise be scrollable.