feat: Plumbs content hash to post-submit LUCI builds (#4795)
This change plumbs the content-aware hash from the scheduler to the LUCI
build service for post-submit builds. This will allow LUCI/recipes to
use the content hash in post-submit builds (windows arm, ddm, etc).

The following changes are included:
- The `ContentAwareHashService` is updated to include a
`getHashByCommitSha` method.
- The `Scheduler` now retrieves the content hash on addCommit() and
passes it to `LuciBuildService`.
- `LuciBuildService` plumbs the `contentHash` through to buildbucket
properties.
- Tests have been added to verify that the content hash is correctly
passed to the `LuciBuildService`.

Extension of flutter/flutter#167779