feat(content-aware-hashing): Allow builds to take over failed hashes
fixes: flutter/flutter#172694

When a merge group is destroyed (for dequeued or invalided reasons), the
corresponding content-aware hash build is now explicitly marked as
failed. This short-curcuits waiting for LUCI jobs to be canceled.

When a content-aware hash build fails, it can now be taken over by a subsequent
build for the same content hash. This happens when a PR is removed from
the merge queue and an "engine" PR is added back - the previous failure
will be restarted and "content_hash" passed to LUCI recipes.

The `ContentAwareHashBuilds` model now includes a list of `failedCommitShas` to
track the history of failed attempts for a given content hash. When a new build
takes over a failed one, the old commit SHA is added to this list.